refactor(analysis): externalize point2 and point3 inline tests
