+++ foundation/analysis/src/linalg/mod.rs
@@ -18,6 +18,10 @@ pub mod vector;
 
 // テストモジュール（*_tests.rs形式）
 #[cfg(test)]
+pub mod point2_tests;
+#[cfg(test)]
+pub mod point3_tests;
+#[cfg(test)]
 pub mod quaternion_tests;
 #[cfg(test)]
 pub mod scalar_tests;
